+++ src/Client.php
@@ -19,10 +19,8 @@ use Bugsnag\Request\ResolverInterface;
 use Bugsnag\Shutdown\PhpShutdownStrategy;
 use Bugsnag\Shutdown\ShutdownStrategyInterface;
 use Composer\CaBundle\CaBundle;
-use GuzzleHttp\Client as Guzzle;
+use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\ClientInterface;
-use ReflectionClass;
-use ReflectionException;
 
 class Client
 {
@@ -149,7 +147,7 @@ class Client
             $options['verify'] = $path;
         }
 
-        return new Guzzle($options);
+        return new GuzzleClient($options);
     }
 
     /**
@@ -242,12 +240,6 @@ class Client
      */
     public function leaveBreadcrumb($name, $type = null, array $metaData = [])
     {
-        try {
-            $name = (new ReflectionClass($name))->getShortName();
-        } catch (ReflectionException $e) {
-            //
-        }
-
         $type = in_array($type, Breadcrumb::getTypes(), true) ? $type : Breadcrumb::MANUAL_TYPE;
 
         $this->recorder->record(new Breadcrumb($name, $type, $metaData));
@@ -806,7 +798,7 @@ class Client
     /**
      * Get the session client.
      *
-     * @return \Guzzle\ClientInterface
+     * @return \GuzzleHttp\ClientInterface
      */
     public function getSessionClient()
     {