+++ src/SessionTracker.php
@@ -2,6 +2,10 @@
 
 namespace Bugsnag;
 
+use Exception;
+use GuzzleHttp\ClientInterface;
+use InvalidArgumentException;
+
 class SessionTracker
 {
     /**
@@ -105,14 +109,23 @@ class SessionTracker
         $this->lastSent = 0;
     }
 
+    /**
+     * @param Configuration $config
+     *
+     * @return void
+     */
     public function setConfig(Configuration $config)
     {
         $this->config = $config;
     }
 
+    /**
+     * @return void
+     */
     public function startSession()
     {
         $currentTime = strftime('%Y-%m-%dT%H:%M:00');
+
         $session = [
             'id' => uniqid('', true),
             'startedAt' => $currentTime,
@@ -121,32 +134,50 @@ class SessionTracker
                 'unhandled' => 0,
             ],
         ];
+
         $this->setCurrentSession($session);
         $this->incrementSessions($currentTime);
     }
 
+    /**
+     * @param array $session
+     *
+     * @return void
+     */
     protected function setCurrentSession(array $session)
     {
-        if (!is_null($this->sessionFunction)) {
+        if (is_callable($this->sessionFunction)) {
             call_user_func($this->sessionFunction, $session);
         } else {
             $this->currentSession = $session;
         }
     }
 
+    /**
+     * @return array
+     */
     public function getCurrentSession()
     {
-        if (!is_null($this->sessionFunction)) {
-            return call_user_func($this->sessionFunction);
-        } else {
-            return $this->currentSession;
+        if (is_callable($this->sessionFunction)) {
+            $currentSession = call_user_func($this->sessionFunction);
+
+            if (is_array($currentSession)) {
+                return $currentSession;
+            }
+
+            return [];
         }
+
+        return $this->currentSession;
     }
 
+    /**
+     * @return void
+     */
     public function sendSessions()
     {
         $locked = false;
-        if (!is_null($this->lockFunction) && !is_null($this->unlockFunction)) {
+        if (is_callable($this->lockFunction) && is_callable($this->unlockFunction)) {
             call_user_func($this->lockFunction);
             $locked = true;
         }
@@ -160,57 +191,82 @@ class SessionTracker
         }
     }
 
+    /**
+     * @param callable $lock
+     * @param callable $unlock
+     *
+     * @return void
+     */
     public function setLockFunctions($lock, $unlock)
     {
-        if (is_callable($lock) && is_callable($unlock)) {
-            $this->lockFunction = $lock;
-            $this->unlockFunction = $unlock;
-        } else {
+        if (!is_callable($lock) || !is_callable($unlock)) {
             throw new InvalidArgumentException('Both lock and unlock functions must be callable');
         }
+
+        $this->lockFunction = $lock;
+        $this->unlockFunction = $unlock;
     }
 
-    public function setRetryFunction($retry)
+    /**
+     * @param callable $function
+     *
+     * @return void
+     */
+    public function setRetryFunction($function)
     {
-        if (is_callable($retry)) {
-            $this->retryFunction = $retry;
-        } else {
+        if (!is_callable($function)) {
             throw new InvalidArgumentException('The retry function must be callable');
         }
+
+        $this->retryFunction = $function;
     }
 
+    /**
+     * @param callable $function
+     *
+     * @return void
+     */
     public function setStorageFunction($function)
     {
-        if (is_callable($function)) {
-            $this->storageFunction = $function;
-        } else {
+        if (!is_callable($function)) {
             throw new InvalidArgumentException('Storage function must be callable');
         }
+
+        $this->storageFunction = $function;
     }
 
+    /**
+     * @param callable $function
+     *
+     * @return void
+     */
     public function setSessionFunction($function)
     {
-        if (is_callable($function)) {
-            $this->sessionFunction = $function;
-        } else {
+        if (!is_callable($function)) {
             throw new InvalidArgumentException('Session function must be callable');
         }
+
+        $this->sessionFunction = $function;
     }
 
+    /**
+     * @param string $minute
+     * @param int $count
+     * @param bool $deliver
+     *
+     * @return void
+     */
     protected function incrementSessions($minute, $count = 1, $deliver = true)
     {
         $locked = false;
-        if (!is_null($this->lockFunction) && !is_null($this->unlockFunction)) {
+
+        if (is_callable($this->lockFunction) && is_callable($this->unlockFunction)) {
             call_user_func($this->lockFunction);
             $locked = true;
         }
 
         try {
             $sessionCounts = $this->getSessionCounts();
-
-            if (is_null($sessionCounts)) {
-                $sessionCounts = [];
-            }
 
             if (array_key_exists($minute, $sessionCounts)) {
                 $sessionCounts[$minute] += $count;
@@ -223,7 +279,9 @@ class SessionTracker
             if (count($sessionCounts) > self::$MAX_SESSION_COUNT) {
                 $this->trimOldestSessions();
             }
+
             $lastSent = $this->getLastSent();
+
             if ($deliver && ((time() - $lastSent) > self::$DELIVERY_INTERVAL)) {
                 $this->deliverSessions();
             }
@@ -234,36 +292,61 @@ class SessionTracker
         }
     }
 
+    /**
+     * @return array
+     */
     protected function getSessionCounts()
     {
-        if (!is_null($this->storageFunction)) {
-            return call_user_func($this->storageFunction, self::$SESSION_COUNTS_KEY);
-        } else {
-            return $this->sessionCounts;
+        if (is_callable($this->storageFunction)) {
+            $sessionCounts = call_user_func($this->storageFunction, self::$SESSION_COUNTS_KEY);
+
+            if (is_array($sessionCounts)) {
+                return $sessionCounts;
+            }
+
+            return [];
         }
+
+        return $this->sessionCounts;
     }
 
+    /**
+     * @param array $sessionCounts
+     *
+     * @return void
+     */
     protected function setSessionCounts(array $sessionCounts)
     {
-        if (!is_null($this->storageFunction)) {
-            return call_user_func($this->storageFunction, self::$SESSION_COUNTS_KEY, $sessionCounts);
-        } else {
-            $this->sessionCounts = $sessionCounts;
+        if (is_callable($this->storageFunction)) {
+            call_user_func($this->storageFunction, self::$SESSION_COUNTS_KEY, $sessionCounts);
         }
+
+        $this->sessionCounts = $sessionCounts;
     }
 
+    /**
+     * @return void
+     */
     protected function trimOldestSessions()
     {
         $sessions = $this->getSessionCounts();
+
         uksort($sessions, function ($key) {
             return strtotime($key);
         });
+
         $sessions = array_reverse($sessions);
         $sessionCounts = array_slice($sessions, 0, self::$MAX_SESSION_COUNT);
+
         $this->setSessionCounts($sessionCounts);
     }
 
-    protected function constructPayload($sessions)
+    /**
+     * @param array $sessions
+     *
+     * @return array
+     */
+    protected function constructPayload(array $sessions)
     {
         $formattedSessions = [];
         foreach ($sessions as $minute => $count) {
@@ -278,33 +361,48 @@ class SessionTracker
         ];
     }
 
+    /**
+     * @return void
+     */
     protected function deliverSessions()
     {
         $sessions = $this->getSessionCounts();
+
         $this->setSessionCounts([]);
-        if (count($sessions) == 0) {
+
+        if (count($sessions) === 0) {
             return;
         }
+
         if (!$this->config->shouldNotify()) {
             return;
         }
+
         $http = $this->config->getSessionClient();
-        $payload = $this->constructPayload($sessions);
-        $headers = [
-            'Bugsnag-Api-Key' => $this->config->getApiKey(),
-            'Bugsnag-Payload-Version' => self::$SESSION_PAYLOAD_VERSION,
-            'Bugsnag-Sent-At' => strftime('%Y-%m-%dT%H:%M:%S'),
+
+        $options = [
+            'json' => $this->constructPayload($sessions),
+            'headers' => [
+                'Bugsnag-Api-Key' => $this->config->getApiKey(),
+                'Bugsnag-Payload-Version' => self::$SESSION_PAYLOAD_VERSION,
+                'Bugsnag-Sent-At' => strftime('%Y-%m-%dT%H:%M:%S'),
+            ],
         ];
+
         $this->setLastSent();
 
         try {
-            $http->post('', [
-                'json' => $payload,
-                'headers' => $headers,
-            ]);
+            // Support later Guzzle versions — note we check the interface to make
+            // sure "request" is a public method as on PHP 7.4 "method_exists" will
+            // return true for private methods
+            if (method_exists(ClientInterface::class, 'request')) {
+                $http->request('POST', '', $options);
+            } else {
+                $http->post('', $options);
+            }
         } catch (Exception $e) {
             error_log('Bugsnag Warning: Couldn\'t notify. '.$e->getMessage());
-            if (!is_null($this->retryFunction)) {
+            if (is_callable($this->retryFunction)) {
                 call_user_func($this->retryFunction, $sessions);
             } else {
                 foreach ($sessions as $minute => $count) {
@@ -314,22 +412,35 @@ class SessionTracker
         }
     }
 
+    /**
+     * @return void
+     */
     protected function setLastSent()
     {
         $time = time();
-        if (!is_null($this->storageFunction)) {
+
+        if (is_callable($this->storageFunction)) {
             call_user_func($this->storageFunction, self::$SESSIONS_LAST_SENT_KEY, $time);
         } else {
             $this->lastSent = $time;
         }
     }
 
+    /**
+     * @return int
+     */
     protected function getLastSent()
     {
-        if (!is_null($this->storageFunction)) {
-            return call_user_func($this->storageFunction, self::$SESSIONS_LAST_SENT_KEY);
-        } else {
-            return $this->lastSent;
+        if (is_callable($this->storageFunction)) {
+            $lastSent = call_user_func($this->storageFunction, self::$SESSIONS_LAST_SENT_KEY);
+
+            if (is_int($lastSent)) {
+                return $lastSent;
+            }
+
+            return 0;
         }
+
+        return $this->lastSent;
     }
 }