+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+## 3.21.0 (2020-04-29)
+
+### Enhancements
+
+* Use full class names in breadcrumbs
+  [#578](https://github.com/bugsnag/bugsnag-php/pull/578)
+
+* Added support for getting the "original error"
+  [#576](https://github.com/bugsnag/bugsnag-php/pull/576)
+
+### Bug Fixes
+
+* Fixed session tracker possibly raising a warning on PHP 7.2 and above
+  [#579](https://github.com/bugsnag/bugsnag-php/pull/579)
+
 ## 3.20.0 (2020-02-26)
 
 ### Enhancements